+++ src/main/java/com/tcc/CadeMeuBichinho/repository/UserRepository.java
@@ -7,5 +7,5 @@ import com.tcc.CadeMeuBichinho.model.User;
 public interface UserRepository extends CrudRepository<User, Long> {
 	public User findByEmailAndPassword(String email,String password);
 	public User findByEmail(String email);
-	public User findByEmailAndPhone(String email,Integer phone);
+	public User findByEmailAndPhone(String email,Long phone);
 }
